+++ lua/plugins/render-markdown.lua
@@ -1,0 +1,13 @@
+-- lua/plugins/render-markdown.lua
+
+return {
+    'MeanderingProgrammer/render-markdown.nvim',
+    ft           = { 'markdown' },
+    dependencies = {
+        'nvim-treesitter/nvim-treesitter',
+        'nvim-tree/nvim-web-devicons',
+    },
+    config = function()
+        require('config.render-markdown')
+    end,
+}